added MacOS suport

+++ samples/OneClickDemo/OneClickDemo-Katalog.sh
@@ -159,18 +159,21 @@ if [ ${USE_KIND} -eq 1 ]; then
     fi
 fi
 
-header "Checking for bin/7zzs"
-if [[ -f bin/7zzs ]]
+if [[ "$os" != "darwin" ]]
 then
-    header "  7z already exists"
-else 
-    header "Installing bin/7zzs"
-    mkdir -p 7z-install
-    curl -L https://www.7-zip.org/a/7z2201-linux-x64.tar.xz -o 7z-install/7z.tar.xz
-    tar -xf 7z-install/7z.tar.xz -C 7z-install
-    chmod u+x 7z-install/7zzs
-    mv 7z-install/7zzs ./bin
-    rm -r 7z-install
+  header "Checking for bin/7zzs"
+  if [[ -f bin/7zzs ]]
+  then
+      header "  7z already exists"
+  else 
+      header "Installing bin/7zzs"
+      mkdir -p 7z-install
+      curl -L https://www.7-zip.org/a/7z2201-linux-x64.tar.xz -o 7z-install/7z.tar.xz
+      tar -xf 7z-install/7z.tar.xz -C 7z-install
+      chmod u+x 7z-install/7zzs
+      mv 7z-install/7zzs ./bin
+      rm -r 7z-install
+  fi
 fi
 
 header "Checking for aws-cli v2"
@@ -178,15 +181,42 @@ if [[ -f bin/aws && -d bin/aws-source/v2 ]]
 then
     header "  bin/aws v2 already exists"
 else
-    header "Installing bin/aws ${AWSCLI_VERSION}" 
-    # Installed this way due to a known open bug: https://github.com/aws/aws-cli/issues/6852
-    mkdir -p awscli-install
-    curl "https://awscli.amazonaws.com/awscli-exe-linux-x86_64-${AWSCLI_VERSION}.zip" -o awscli-install/awscliv2.zip
-    ./bin/7zzs x awscli-install/awscliv2.zip -oawscli-install
-    ./awscli-install/aws/install -i bin/aws-source -b bin
-    rm bin/aws bin/aws_completer
-    ln -s aws-source/v2/${AWSCLI_VERSION}/bin/aws ./bin/aws
-    rm -r ./awscli-install
+    header "Installing bin/aws" 
+    if [[ "$os" == "darwin" ]]
+    then
+      echo "<?xml version=\"1.0\" encoding=\"UTF-8\"?>
+<"'!'"DOCTYPE plist PUBLIC \"-//Apple//DTD PLIST 1.0//EN\" \"http://www.apple.com/DTDs/PropertyList-1.0.dtd\">
+<plist version=\"1.0\">
+  <array>
+    <dict>
+      <key>choiceAttribute</key>
+      <string>customLocation</string>
+      <key>attributeSetting</key>
+      <string>/$(pwd)/bin</string>
+      <key>choiceIdentifier</key>
+      <string>default</string>
+    </dict>
+  </array>
+</plist>" > bin/choices.xml
+
+      curl "https://awscli.amazonaws.com/AWSCLIV2.pkg" -o "AWSCLIV2.pkg"
+      installer -pkg AWSCLIV2.pkg \
+              -target CurrentUserHomeDirectory \
+              -applyChoiceChangesXML bin/choices.xml
+      ln -s /bin/aws-cli/aws /bin/aws
+      ln -s /bin/aws-cli/aws_completer /bin/aws_completer
+      rm AWSCLIV2.pkg
+
+    else
+      # Installed this way due to a known open bug: https://github.com/aws/aws-cli/issues/6852
+      mkdir -p awscli-install
+      curl "https://awscli.amazonaws.com/awscli-exe-linux-x86_64-${AWSCLI_VERSION}.zip" -o awscli-install/awscliv2.zip
+      ./bin/7zzs x awscli-install/awscliv2.zip -oawscli-install
+      ./awscli-install/aws/install -i bin/aws-source -b bin
+      rm bin/aws bin/aws_completer
+      ln -s aws-source/v2/${AWSCLI_VERSION}/bin/aws ./bin/aws
+      rm -r ./awscli-install
+    fi
 fi
